+++ clients/privacy-center/features/common/config.slice.ts
@@ -2,28 +2,13 @@ import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 
 import { useAppSelector } from "~/app/hooks";
 import type { RootState } from "~/app/store";
-import { getDefaultConfig } from "~/constants";
 import { Consent, ConsentPreferences } from "~/types/api";
 import { Config } from "~/types/config";
 
-// DEFER: by statically loading the config here, we *guarantee* that the initial
-// state is never undefined. This is convenient for all the Typescript checking,
-// but is probably not wise - do we *really* want to ever show the "default"
-// prebuilt config?
-//
-// We should refactor this to initialize the slice when the store is initialized
-// (at runtime) and then set the initial state then. That would allow us to define the type as:
-// ```
-// type State = Config | undefined
-// ```
-//
-// ...and then update all the code to handle that.
-//
-// (see https://github.com/ethyca/fides/issues/3212)
 interface ConfigState {
-  config: Config;
+  config?: Config;
 }
-const initialState: ConfigState = { config: getDefaultConfig() };
+const initialState: ConfigState = {};
 
 export const configSlice = createSlice({
   name: "config",
@@ -32,13 +17,25 @@ export const configSlice = createSlice({
     /**
      * Load a new configuration, replacing the current state entirely.
      */
-    loadConfig(draftState, { payload }: PayloadAction<Config>) {
+    loadConfig(draftState, { payload }: PayloadAction<Config | undefined>) {
+      if (process.env.NODE_ENV === "development") {
+        // eslint-disable-next-line no-console
+        console.log(
+          "Loading Privacy Center configuration into Redux store...",
+          payload?.title
+        );
+      }
       draftState.config = payload;
     },
     /**
      * Modify the current configuration, by merging a partial config into the current state.
      */
     mergeConfig(draftState, { payload }: PayloadAction<Partial<Config>>) {
+      if (!draftState.config) {
+        throw new Error(
+          "Cannot apply mergeConfig into uninitialized Redux state; must use loadConfig first!"
+        );
+      }
       draftState.config = { ...draftState.config, ...payload };
     },
     /**
@@ -49,6 +46,11 @@ export const configSlice = createSlice({
       draftState,
       { payload }: PayloadAction<ConsentPreferences>
     ) {
+      if (!draftState.config) {
+        throw new Error(
+          "Cannot apply updateConsentOptionsFromApi into uninitialized Redux state; must use loadConfig first!"
+        );
+      }
       const consentPreferencesMap = new Map<string, Consent>(
         (payload.consent ?? []).map((consent) => [consent.data_use, consent])
       );
@@ -74,4 +76,10 @@ const selectConfig = (state: RootState) => state.config;
 export const { reducer } = configSlice;
 export const { loadConfig, mergeConfig, updateConsentOptionsFromApi } =
   configSlice.actions;
-export const useConfig = (): Config => useAppSelector(selectConfig).config;
+export const useConfig = (): Config => {
+  const { config } = useAppSelector(selectConfig);
+  if (!config) {
+    throw new Error("Unable to load Privacy Center configuration!");
+  }
+  return config;
+};