+++ CHANGELOG.md
@@ -31,6 +31,9 @@ The types of changes are:
 - Restrict where privacy notices of certain consent mechanisms must be displayed [#3195](https://github.com/ethyca/fides/pull/3195)
 - Merged the `lib` submodule into the `api.ops` submodule [#3134](https://github.com/ethyca/fides/pull/3134)
 
+### Fixed
+- Prevent ability to unintentionally show "default" Privacy Center configuration, styles, etc. [#3242](https://github.com/ethyca/fides/pull/3242)
+
 ### Developer Experience
 
 - Use prettier to format *all* source files in client packages [#3240](https://github.com/ethyca/fides/pull/3240)